snapshot chunk and restore traits
