+++ ethcore/src/snapshot/snapshot_service_trait.rs
@@ -48,10 +48,6 @@ pub trait SnapshotService : Sync + Send {
 	/// Feed a raw block chunk to the service to be processed asynchronously.
 	/// no-op if currently restoring.
 	fn restore_block_chunk(&self, hash: H256, chunk: Bytes);
-
-	/// Give the restoration in-progress some canonical block hashes for
-	/// extra verification (performed at the end)
-	fn provide_canon_hashes(&self, canonical: &[(u64, H256)]);
 }
 
 impl IpcConfig for SnapshotService { }
