+++ ethcore/src/snapshot/tests/blocks.rs
@@ -21,13 +21,12 @@ use error::Error;
 
 use blockchain::generator::{ChainGenerator, ChainIterator, BlockFinalizer};
 use blockchain::BlockChain;
-use snapshot::{chunk_blocks, BlockRebuilder, Error as SnapshotError, Progress};
+use snapshot::{chunk_secondary, Error as SnapshotError, Progress, SnapshotComponents};
 use snapshot::io::{PackedReader, PackedWriter, SnapshotReader, SnapshotWriter};
 
 use util::{Mutex, snappy};
-use util::kvdb::{Database, DatabaseConfig};
+use util::kvdb::{self, KeyValueDB, DBTransaction};
 
-use std::collections::HashMap;
 use std::sync::Arc;
 use std::sync::atomic::AtomicBool;
 
@@ -35,19 +34,18 @@ fn chunk_and_restore(amount: u64) {
 	let mut canon_chain = ChainGenerator::default();
 	let mut finalizer = BlockFinalizer::default();
 	let genesis = canon_chain.generate(&mut finalizer).unwrap();
-	let db_cfg = DatabaseConfig::with_columns(::db::NUM_COLUMNS);
+	let components = ::snapshot::PowSnapshot;
 
 	let engine = Arc::new(::engines::NullEngine::default());
-	let orig_path = RandomTempPath::create_dir();
 	let new_path = RandomTempPath::create_dir();
 	let mut snapshot_path = new_path.as_path().to_owned();
 	snapshot_path.push("SNAP");
 
-	let old_db = Arc::new(Database::open(&db_cfg, orig_path.as_str()).unwrap());
+	let old_db = Arc::new(kvdb::in_memory(::db::NUM_COLUMNS.unwrap_or(0)));
 	let bc = BlockChain::new(Default::default(), &genesis, old_db.clone());
 
 	// build the blockchain.
-	let mut batch = old_db.transaction();
+	let mut batch = DBTransaction::new();
 	for _ in 0..amount {
 		let block = canon_chain.generate(&mut finalizer).unwrap();
 		bc.insert_block(&mut batch, &block, vec![]);
@@ -56,12 +54,18 @@ fn chunk_and_restore(amount: u64) {
 
 	old_db.write(batch).unwrap();
 
-
 	let best_hash = bc.best_block_hash();
 
 	// snapshot it.
 	let writer = Mutex::new(PackedWriter::new(&snapshot_path).unwrap());
-	let block_hashes = chunk_blocks(&bc, best_hash, &writer, &Progress::default()).unwrap();
+	let block_hashes = chunk_secondary(
+		Box::new(::snapshot::PowSnapshot),
+		&bc,
+		best_hash,
+		&writer,
+		&Progress::default()
+	).unwrap();
+
 	let manifest = ::snapshot::ManifestData {
 		version: 2,
 		state_hashes: Vec::new(),
@@ -74,9 +78,10 @@ fn chunk_and_restore(amount: u64) {
 	writer.into_inner().finish(manifest.clone()).unwrap();
 
 	// restore it.
-	let new_db = Arc::new(Database::open(&db_cfg, new_path.as_str()).unwrap());
+	let new_db = Arc::new(kvdb::in_memory(::db::NUM_COLUMNS.unwrap_or(0)));
 	let new_chain = BlockChain::new(Default::default(), &genesis, new_db.clone());
-	let mut rebuilder = BlockRebuilder::new(new_chain, new_db.clone(), &manifest).unwrap();
+	let mut rebuilder = components.rebuilder(new_chain, new_db.clone(), &manifest).unwrap();
+
 	let reader = PackedReader::new(&snapshot_path).unwrap().unwrap();
 	let flag = AtomicBool::new(true);
 	for chunk_hash in &reader.manifest().block_hashes {
@@ -85,7 +90,8 @@ fn chunk_and_restore(amount: u64) {
 		rebuilder.feed(&chunk, engine.as_ref(), &flag).unwrap();
 	}
 
-	rebuilder.finalize(HashMap::new()).unwrap();
+	rebuilder.finalize().unwrap();
+	drop(rebuilder);
 
 	// and test it.
 	let new_chain = BlockChain::new(Default::default(), &genesis, new_db);
@@ -118,10 +124,8 @@ fn checks_flag() {
 	};
 
 	let chunk = stream.out();
-	let path = RandomTempPath::create_dir();
 
-	let db_cfg = DatabaseConfig::with_columns(::db::NUM_COLUMNS);
-	let db = Arc::new(Database::open(&db_cfg, path.as_str()).unwrap());
+	let db = Arc::new(kvdb::in_memory(::db::NUM_COLUMNS.unwrap_or(0)));
 	let engine = Arc::new(::engines::NullEngine::default());
 	let chain = BlockChain::new(Default::default(), &genesis, db.clone());
 
@@ -134,7 +138,7 @@ fn checks_flag() {
 		block_hash: H256::default(),
 	};
 
-	let mut rebuilder = BlockRebuilder::new(chain, db.clone(), &manifest).unwrap();
+	let mut rebuilder = ::snapshot::PowSnapshot.rebuilder(chain, db.clone(), &manifest).unwrap();
 
 	match rebuilder.feed(&chunk, engine.as_ref(), &AtomicBool::new(false)) {
 		Err(Error::Snapshot(SnapshotError::RestorationAborted)) => {}
