+++ rpc/src/v1/tests/helpers/snapshot_service.rs
@@ -47,5 +47,4 @@ impl SnapshotService for TestSnapshotService {
 	fn abort_restore(&self) { }
 	fn restore_state_chunk(&self, _hash: H256, _chunk: Bytes) { }
 	fn restore_block_chunk(&self, _hash: H256, _chunk: Bytes) { }
-	fn provide_canon_hashes(&self, _hashes: &[(u64, H256)]) { }
 }
